requirements.txt
robotframework
robotframework-requests
RESTinstance
robotframework-jsonlibrary
robotframework-restlibrary
python-dotenv
httpx
playwright